Added npm package steam, basic bot login
Setup the basic functions of the trade bot in app.js. The npm package Steam provides methods for logging in to steam taking username and password from environment variables. If the account is protected by steam guard, which uses two step authentication via email code, the first login on a new machine will fail. Retreive the code from email and set to logOnOptions authCode, the machine will generate a sentryFile data blob that is used to log in from now on, you can remove the authCode.